Separate init and processing data for sample scatter and single cell scatter
